new Updates

+++ app/Http/Controllers/CourseController.php
@@ -41,6 +41,10 @@ class CourseController extends Controller
     public function delete($id)
     {
         $course = Course::where('id', $id)->first();
+        if($course->invs()->count() > 0)
+        {
+            return response(['message' => 'Cannot delete the course there are invs attached to it.'],402);
+        }
         Course::destroy($id);
         return response(['message' => $course->code.' - '.$course->name.' Removed successfully']);
     }
